Move applyFilter, selectRange, valueClick triggers to data/embeddables.
