+++ src/plugins/ui_actions/public/index.ts
@@ -40,12 +40,6 @@ export {
 export {
   Trigger,
   TriggerContext,
-  SELECT_RANGE_TRIGGER,
-  selectRangeTrigger,
-  VALUE_CLICK_TRIGGER,
-  valueClickTrigger,
-  APPLY_FILTER_TRIGGER,
-  applyFilterTrigger,
   VISUALIZE_FIELD_TRIGGER,
   visualizeFieldTrigger,
   VISUALIZE_GEO_FIELD_TRIGGER,
