+++ src/plugins/ui_actions/public/plugin.ts
@@ -20,14 +20,7 @@
 import { CoreStart, CoreSetup, Plugin, PluginInitializerContext } from 'src/core/public';
 import { PublicMethodsOf } from '@kbn/utility-types';
 import { UiActionsService } from './service';
-import {
-  selectRangeTrigger,
-  valueClickTrigger,
-  rowClickTrigger,
-  applyFilterTrigger,
-  visualizeFieldTrigger,
-  visualizeGeoFieldTrigger,
-} from './triggers';
+import { rowClickTrigger, visualizeFieldTrigger, visualizeGeoFieldTrigger } from './triggers';
 
 export type UiActionsSetup = Pick<
   UiActionsService,
@@ -47,10 +40,7 @@ export class UiActionsPlugin implements Plugin<UiActionsSetup, UiActionsStart> {
   constructor(initializerContext: PluginInitializerContext) {}
 
   public setup(core: CoreSetup): UiActionsSetup {
-    this.service.registerTrigger(selectRangeTrigger);
-    this.service.registerTrigger(valueClickTrigger);
     this.service.registerTrigger(rowClickTrigger);
-    this.service.registerTrigger(applyFilterTrigger);
     this.service.registerTrigger(visualizeFieldTrigger);
     this.service.registerTrigger(visualizeGeoFieldTrigger);
     return this.service;
