+++ src/plugins/ui_actions/public/triggers/index.ts
@@ -20,10 +20,7 @@
 export * from './trigger';
 export * from './trigger_contract';
 export * from './trigger_internal';
-export * from './select_range_trigger';
-export * from './value_click_trigger';
 export * from './row_click_trigger';
-export * from './apply_filter_trigger';
 export * from './visualize_field_trigger';
 export * from './visualize_geo_field_trigger';
 export * from './default_trigger';
