+++ src/plugins/ui_actions/public/triggers/trigger_contract.ts
@@ -25,8 +25,7 @@ import { TriggerId, TriggerContextMapping } from '../types';
  */
 export class TriggerContract<T extends TriggerId> {
   /**
-   * Unique name of the trigger as identified in `ui_actions` plugin trigger
-   * registry, such as "SELECT_RANGE_TRIGGER" or "VALUE_CLICK_TRIGGER".
+   * Unique name of the trigger as identified in `ui_actions` plugin trigger registry.
    */
   public readonly id: T;
