+++ src/plugins/ui_actions/public/types.ts
@@ -20,17 +20,12 @@
 import { ActionInternal } from './actions/action_internal';
 import { TriggerInternal } from './triggers/trigger_internal';
 import {
-  SELECT_RANGE_TRIGGER,
-  VALUE_CLICK_TRIGGER,
   ROW_CLICK_TRIGGER,
-  APPLY_FILTER_TRIGGER,
   VISUALIZE_FIELD_TRIGGER,
   VISUALIZE_GEO_FIELD_TRIGGER,
   DEFAULT_TRIGGER,
   RowClickContext,
 } from './triggers';
-import type { RangeSelectContext, ValueClickContext } from '../../embeddable/public';
-import type { ApplyGlobalFilterActionContext } from '../../data/public';
 
 export type TriggerRegistry = Map<TriggerId, TriggerInternal<any>>;
 export type ActionRegistry = Map<string, ActionInternal>;
@@ -49,10 +44,7 @@ export type TriggerContext = BaseContext;
 
 export interface TriggerContextMapping {
   [DEFAULT_TRIGGER]: TriggerContext;
-  [SELECT_RANGE_TRIGGER]: RangeSelectContext;
-  [VALUE_CLICK_TRIGGER]: ValueClickContext;
   [ROW_CLICK_TRIGGER]: RowClickContext;
-  [APPLY_FILTER_TRIGGER]: ApplyGlobalFilterActionContext;
   [VISUALIZE_FIELD_TRIGGER]: VisualizeFieldContext;
   [VISUALIZE_GEO_FIELD_TRIGGER]: VisualizeFieldContext;
 }
