+++ x-pack/plugins/translations/translations/ja-JP.json
@@ -3490,12 +3490,12 @@
     "uiActions.actionPanel.more": "詳細",
     "uiActions.actionPanel.title": "オプション",
     "uiActions.errors.incompatibleAction": "操作に互換性がありません",
-    "uiActions.triggers.applyFilterDescription": "Kibanaフィルターが適用されるとき。単一の値または範囲フィルターにすることができます。",
-    "uiActions.triggers.applyFilterTitle": "フィルターを適用",
-    "uiActions.triggers.selectRangeDescription": "ビジュアライゼーションでの値の範囲",
-    "uiActions.triggers.selectRangeTitle": "範囲選択",
-    "uiActions.triggers.valueClickDescription": "ビジュアライゼーションでデータポイントをクリック",
-    "uiActions.triggers.valueClickTitle": "シングルクリック",
+    "data.triggers.applyFilterDescription": "Kibanaフィルターが適用されるとき。単一の値または範囲フィルターにすることができます。",
+    "data.triggers.applyFilterTitle": "フィルターを適用",
+    "embeddableApi.selectRangeTrigger.description": "ビジュアライゼーションでの値の範囲",
+    "embeddableApi.selectRangeTrigger.title": "範囲選択",
+    "embeddableApi.valueClickTrigger.description": "ビジュアライゼーションでデータポイントをクリック",
+    "embeddableApi.valueClickTrigger.title": "シングルクリック",
     "usageCollection.stats.notReadyMessage": "まだ統計が準備できていません。しばらくたってから再試行してください。",
     "visDefaultEditor.advancedToggle.advancedLinkLabel": "高度な設定",
     "visDefaultEditor.agg.toggleEditorButtonAriaLabel": "{schema} エディターを切り替える",
