+++ x-pack/plugins/translations/translations/zh-CN.json
@@ -3491,12 +3491,12 @@
     "uiActions.actionPanel.more": "更多",
     "uiActions.actionPanel.title": "选项",
     "uiActions.errors.incompatibleAction": "操作不兼容",
-    "uiActions.triggers.applyFilterDescription": "应用 kibana 筛选时。可能是单个值或范围筛选。",
-    "uiActions.triggers.applyFilterTitle": "应用筛选",
-    "uiActions.triggers.selectRangeDescription": "可视化上的一组值",
-    "uiActions.triggers.selectRangeTitle": "范围选择",
-    "uiActions.triggers.valueClickDescription": "可视化上的数据点单击",
-    "uiActions.triggers.valueClickTitle": "单击",
+    "data.triggers.applyFilterDescription": "应用 kibana 筛选时。可能是单个值或范围筛选。",
+    "data.triggers.applyFilterTitle": "应用筛选",
+    "embeddableApi.selectRangeTrigger.description": "可视化上的一组值",
+    "embeddableApi.selectRangeTrigger.title": "范围选择",
+    "embeddableApi.valueClickTrigger.description": "可视化上的数据点单击",
+    "embeddableApi.valueClickTrigger.title": "单击",
     "usageCollection.stats.notReadyMessage": "统计信息尚未准备就绪。请稍后重试。",
     "visDefaultEditor.advancedToggle.advancedLinkLabel": "高级",
     "visDefaultEditor.agg.toggleEditorButtonAriaLabel": "切换 {schema} 编辑器",
